Update Navigation Styling and Event Year
The navigation section has been styled and made more flexible for a better user experience. The layout of the navigation bar has been changed to incorporate a more responsive and intuitive design. Elements within the navigation bar have been styled for better visual hierarchy. The year of the event in the main page and the layout title has been updated to 2024. In addition, the units for radius in the color styles have been changed from pixel to rem for better responsiveness.