Add call for papers section and update dependencies
The new Call for Papers section, represented by the C4P component, has been added to the main page. This includes the addition of the Divider component to distinguish between sections. Adjustments have been made to the UI component stylings for improved UX. In addition, the astro-icon dependency has been added to the project for iconographic displays.