Revamp Sponsors section in event page
The Sponsors section was improved for better visual layout and message delivery. The style was updated to make it responsive, with changes to 'flex' settings helping to maintain structure on different screen sizes. A short introduction and call-to-action for potential sponsors were added to showcase the value of our event and encourage sponsorship. The adjustments enhance the user experience across various devices, especially on mobile, and make information about the event sponsorship clearer and more inviting.